backend/requirements.txt
# Dependências do Gestão de Pacientes

# Framework web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Banco de dados
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9  # Driver PostgreSQL para Neon

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Criptografia para CPF
cryptography==41.0.7

# Validação de dados
pydantic>=2.10.0
pydantic[email]>=2.10.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# Rate limiting
slowapi==0.1.9

# Manipulação de Excel
pandas>=2.2.3
openpyxl>=3.1.5

# CORS
# (incluído no FastAPI)

